requirements.txt
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
alembic==1.10.2
ansible==9.4.0
ansible-core==2.16.5
anyio==3.6.2
apache-airflow==2.5.2
apache-airflow-providers-common-sql==1.3.4
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.2.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-sqlite==3.3.1
apispec==3.3.2
apturl==0.5.2
argcomplete==3.0.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
Automat==22.10.0
autopep8==1.7.0
Babel==2.12.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
bidict==0.23.1
bleach==6.0.0
blinker==1.4
branca==0.7.0
Brlapi==0.8.3
cachelib==0.9.0
capstone==4.0.2
cattrs==22.2.0
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
clickclick==20.10.2
cloudpickle==3.0.0
colorama==0.4.4
colorlog==4.8.0
comm==0.1.2
command-not-found==0.3
configobj==5.0.8
ConfigUpdater==3.1.1
connexion==2.14.2
constantly==15.1.0
cron-descriptor==1.2.35
croniter==1.3.8
cryptography==38.0.4
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.4
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==3.2.8
django-extensions==3.1.3
django-widget-tweaks==1.4.8
dnslib==0.9.23
dnspython==2.3.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.19
duplicity==0.8.21
email-validator==1.3.1
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
fabric==3.0.0
Farama-Notifications==0.0.4
fasteners==0.14.1
fastjsonschema==2.16.3
filelock==3.9.0
Flask==2.2.2
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-MySQLdb==2.0.0
Flask-Session==0.4.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
folium==0.15.0
fonttools==4.37.1
fqdn==1.5.1
frozenlist==1.3.3
future==0.18.2
google==3.0.0
gpg==1.16.0
graphviz==0.20.1
greenlet==2.0.2
gunicorn==20.1.0
gymnasium==0.29.1
h11==0.14.0
html5lib==1.1
httpcore==0.16.3
httpie==3.2.1
httplib2==0.20.2
httpx==0.23.3
hyperlink==21.0.0
idna==3.3
imageio==2.23.0
importlib-metadata==4.6.4
incremental==22.10.0
inflection==0.5.1
invoke==2.0.0
IPy==1.1
ipykernel==6.19.4
ipython==8.8.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.3
jupyter_client==7.4.8
jupyter_core==5.1.2
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.22.0
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
linkify-it-py==2.0.0
lockfile==0.12.2
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.2
mistune==2.0.5
modsimpy==1.1.3
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack-python==0.5.6
multidict==6.0.4
mysqlclient==2.2.0
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
netaddr==0.8.0
NetfilterQueue @ git+https://github.com/kti/python-netfilterqueue@0bb948d2c196f033dbb54aca6bcf24eacb14bf7f
netifaces==0.11.0
networkx==2.8.8
notebook==6.5.4
notebook_shim==0.2.2
ntlm-auth==1.5.0
numpy==1.23.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.0
olefile==0.46
opencv-python==4.7.0.68
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pathspec==0.9.0
patsy==0.5.3
pbr==5.11.1
pefile==2022.5.30
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipenv==2023.4.29
platformdirs==2.6.2
pluggy==1.0.0
prison==0.2.1
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.19.5
psutil==5.9.4
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycodestyle==2.9.1
pycparser==2.21
pycrypto==2.6.1
pycups==2.0.1
Pygments==2.14.0
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pykerberos==1.1.14
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==2.4.7
pypcap==1.3.0
pyperclip==1.8.2
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pytesseract==0.3.10
python-apt==2.4.0+ubuntu3
python-daemon==3.0.1
python-dateutil==2.8.2
python-dbusmock==0.27.5
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-engineio==4.9.0
python-json-logger==2.0.7
python-magic==0.4.27
python-nvd3==0.15.0
python-slugify==8.0.1
python-socketio==5.11.2
pytz==2022.1
pytzdata==2020.1
PyWavelets==1.4.1
pywinrm==0.3.0
pyxdg==0.27
PyYAML==5.4.1
pyzmq==24.0.1
reportlab==3.6.8
requests==2.28.1
requests-kerberos==0.12.0
requests-ntlm==1.1.0
resolvelib==0.8.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.3.2
scapy==2.5.0
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.10.0
screen-resolution-extra==0.0.0
screenkey==1.5
seaborn==0.12.2
SecretStorage==3.3.1
Send2Trash==1.8.0
service-identity==21.1.0
setproctitle==1.3.2
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.47
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.40.0
sqlparse==0.4.3
ssh-import-id==5.11
stable-baselines3==2.2.1
stack-data==0.6.2
statsmodels==0.14.0
stevedore==4.1.1
sympy==1.11.1
systemd-python==234
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.2.0
terminado==0.17.1
text-unidecode==1.3
texttable==1.6.4
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.2.1
tk==0.1.0
toml==0.10.2
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
traitlets==5.8.0
Twisted==22.10.0
typing_extensions==4.4.0
ua-parser==0.16.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
uc-micro-py==1.0.1
ufw==0.36.1
ultralytics==8.2.4
unattended-upgrades==0.1
unicodecsv==0.14.1
uri-template==1.2.0
urllib3==1.26.5
usb-creator==0.3.7
user-agents==2.2.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wadllib==1.3.6
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.2.2
wrapt==1.15.0
wsproto==1.2.0
WTForms==3.0.1
xdg==5
xkit==0.0.0
xmltodict==0.12.0
y-py==0.5.9
yarl==1.8.2
ypy-websocket==0.8.2
zipp==1.0.0
zope.interface==5.5.2
